chore: fixed forge-lint warnings

+++ contracts/nft-redeemer/NFTRedeemer.sol
@@ -280,9 +280,9 @@ contract NFTRedeemer is Ownable, Pausable, ReentrancyGuard {
         (uint256 slot, uint256 bit) = _bitmapLocation(tokenId);
 
         if (value) {
-            _alternatePriceBitmap[slot] |= (1 << bit);
+            _alternatePriceBitmap[slot] |= (uint256(1) << bit);
         } else {
-            _alternatePriceBitmap[slot] &= ~(1 << bit);
+            _alternatePriceBitmap[slot] &= ~(uint256(1) << bit);
         }
     }
 
@@ -315,7 +315,7 @@ contract NFTRedeemer is Ownable, Pausable, ReentrancyGuard {
             uint256 tokenId = tokenIds[i];
 
             (uint256 slot, uint256 bit) = _bitmapLocation(tokenId);
-            uint256 bitMask = (1 << bit);
+            uint256 bitMask = uint256(1) << bit;
 
             // Check if we have already encountered this slot
             bool found = false;
